refactor: removing shorthand that ties properties that require recursive tree structure
