+++ assets/styles/util/ImagePlacement.ts
@@ -1,26 +1,15 @@
 import { Image, ImageStyle } from 'react-native'
-import { Layer, IBaseProps } from './Layer'
 import { ImageAsset } from './ImageAsset'
 import { Placement, IFrameData } from './Placement'
 
-export class ImagePlacement <TParent extends Layer = Layer> {
+export class ImagePlacement {
   place: Placement
   asset: () => ImageAsset
-  parent: TParent
 
-  constructor (asset: () => ImageAsset, frame: IFrameData, parent: TParent) {
+  constructor (asset: () => ImageAsset, frame: IFrameData) {
     this.asset = asset
     this.place = new Placement(frame)
-    this.parent = parent
-    this.Render = this.Render.bind(this)
     this.img = this.img.bind(this)
-  }
-
-  Render (props: IBaseProps<Image, ImageStyle>): JSX.Element {
-    return this.parent.Image({
-      ...props,
-      prototype: this
-    })
   }
 
   img (style?: ImageStyle, ref?: React.Ref<Image>, onLayout?: () => any): JSX.Element {
