+++ assets/styles/util/Layer.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Image, ImageStyle, TextStyle, TextInput, Text, View, ViewStyle, FlexStyle, TouchableOpacity, GestureResponderEvent, Insets, ReturnKeyTypeOptions } from 'react-native'
 import { useVUnits } from './useVUnits'
 import { exists } from './lang'
-import { TextBox } from './TextBox'
+import { TextBox, TTextContentType } from './TextBox'
 import { ImagePlacement } from './ImagePlacement'
 import { Polygon } from './Polygon'
 import { Slice9Placement } from './Slice9Placement'
@@ -36,35 +36,6 @@ export interface IBaseProps<T extends React.Component, TStyle extends FlexStyle>
   onPress?: (event: GestureResponderEvent) => any
   onLayout?: () => any
 }
-
-export type TTextContentType = 'none'
-| 'URL'
-| 'addressCity'
-| 'addressCityAndState'
-| 'addressState'
-| 'countryName'
-| 'creditCardNumber'
-| 'emailAddress'
-| 'familyName'
-| 'fullStreetAddress'
-| 'givenName'
-| 'jobTitle'
-| 'location'
-| 'middleName'
-| 'name'
-| 'namePrefix'
-| 'nameSuffix'
-| 'nickname'
-| 'organizationName'
-| 'postalCode'
-| 'streetAddressLine1'
-| 'streetAddressLine2'
-| 'sublocality'
-| 'telephoneNumber'
-| 'username'
-| 'password'
-| 'newPassword'
-| 'oneTimeCode'
 
 export interface ITextBaseProps extends IBaseProps<Text | TextInput, TextStyle> {
   value?: string
